+++ backend/garpix_page/serializers/__init__.py
@@ -1,2 +1,1 @@
 from .serializer import get_serializer  # noqa
-from .components.components_serializer import get_components_serializer  # noqa
