+++ backend/garpix_page/serializers/components/components_serializer.py
@@ -1,4 +1,3 @@
-from rest_framework.serializers import ModelSerializer
 from rest_polymorphic.serializers import PolymorphicSerializer
 
 from ..serializer import get_serializer
@@ -10,17 +9,6 @@ def get_components_serializers_list():
     for model in get_garpix_page_component_models():
         serializers_list.update({model: get_serializer(model)})
     return serializers_list
-
-
-def get_components_serializer(model):
-    if model.get_serializer(model) is not None:
-        return model.get_serializer(model)
-    return type(f'{model.name}Serializer', (ModelSerializer, ), {
-        'Meta': type('Meta', (object,), {
-            'model': model,
-            'exclude': ('polymorphic_ctype', 'pages')
-        })
-    })
 
 
 class ComponentSerializer(PolymorphicSerializer):
