+++ backend/garpix_page/serializers/serializer.py
@@ -11,3 +11,14 @@ def get_serializer(model):
             # 'fields': '__all__'
         })
     })
+
+
+def get_components_serializer(model):
+    if model.get_serializer(model) is not None:
+        return model.get_serializer(model)
+    return type(f'{model.name}Serializer', (ModelSerializer, ), {
+        'Meta': type('Meta', (object,), {
+            'model': model,
+            'exclude': ('polymorphic_ctype', 'pages')
+        })
+    })
